==> notebooks/01-Read_single_cell.Rmd <==
---
title: 'Reading Steinbock results - non-denoised'
author: "Jacopo Umberto Verga"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

I ran the steinbock pipeline on the non-denoised images, why:
- Odd spillover removal results
- Issues in clustering the cells


# Read in the data 

As suggested in [ImagingWorkshop2023](https://github.com/BodenmillerGroup/ImagingWorkshop2023/tree/main), I am going to use [imcRtools](https://github.com/BodenmillerGroup/imcRtools) to read in single-cell data extracted using the `steinbock` framework. 


```{r, message=FALSE}
library(imcRtools)
library(cytomapper)
library(stringr)
library(scater)
library(BiocParallel)
multicore=MulticoreParam(workers=16)
invisible(BiocManager::install('BiocStyle'))
```

## Read in single-cell information



### steinbock generated data

The downloaded example data (see [data](https://github.com/BodenmillerGroup/ImagingWorkshop2023/tree/main/data)) processed with the
[steinbock](https://github.com/BodenmillerGroup/steinbock) framework can be read
in with the `read_steinbock` function provided by `imcRtools`. For more
information, please refer to `?read_steinbock`.

```{r read-steinbock}
spe <- read_steinbock("../data/non_denoised/", image_file=NULL, BPPARAM = multicore)
# fix rownames to show only gene name:
pattern <- "\\d+([[:alpha:]][[:alpha:]])_([^_]+(?:_[^_]+)?)"
rownames(spe) <- str_match(rownames(spe), pattern)[,3]
spe
```

similar number of cells than with the denoised data, make sense.

```{r counts}
counts(spe)[1:5,1:5]
```

Metadata associated to individual cells are stored in the `colData` slot. After
initial image processing, these metadata include the numeric identifier
(`ObjectNumber`), the area, and morphological features of each cell. In
addition, `sample_id` stores the image name from which each cell was extracted
and the width and height of the corresponding images are stored.

```{r colData}
head(colData(spe))
```

The main difference between the `SpatialExperiment` and the
`SingleCellExperiment` data container in the current setting is the way spatial
locations of all cells are stored. For the `SingleCellExperiment` container, the
locations are stored in the `colData` slot while the `SpatialExperiment`
container stores them in the `spatialCoords` slot:

```{r spatialCoords}
head(spatialCoords(spe))
```

The _spatial cell graphs_ generated by `steinbock` are read into a `colPair`
slot of the `SpatialExperiment` (or `SingleCellExperiment`) object. Cell-cell
interactions (cells in close spatial proximity) are represented as "edge list"
(stored as `SelfHits` object). Here, the left side represents the column indices
of the "from" cells and the right side represents the column indices of the "to"
cells. We will later see how to visualize the _spatial cell graphs_.

```{r colPair}
colPair(spe, "neighborhood")
```

Finally, metadata regarding the channels are stored in the `rowData` slot. This
information is extracted from the `panel.csv` file. Channels are ordered by
isotope mass and therefore match the channel order of the multi-channel images.

```{r rowData}
head(rowData(spe))
```


**NOTE:** I left the deepcell columns to keep track of which channels have been used for the segmentation, all the utilised masks derive from the manual segmentatin with `CellProfiler`


**NOTE2:** I skipped all the section regarding the import of custom files.


## Single-cell processing {#cell-processing}

After reading in the single-cell data, few further processing steps need to be
taken.

**Add additional metadata**

We can set the `colnames` of the object to generate unique identifiers per cell:

```{r set-colnames}
# modified from the original, to keep things easy to read I extracted the tissue identifier and added an s at the beginning
colnames(spe) <- paste0('s',str_split_i(spe$sample_id, "_",2),"_", spe$ObjectNumber)
```

It is also often the case that sample-specific metadata are available externally.
For the current data, we need to link the cancer type (also referred to as "Indication")
to each sample. This metadata is available as external excel file:

```{r add-indication}

meta <- read.csv('../data/non_denoised/meta.csv')

# Add to spe: patient number, timepoint, slide_id, tissue ID

spe$patient_n <- as.factor(meta$Patient_number[match(spe$sample_id,meta$File_name)])
spe$timepoint <- meta$Timepoint[match(spe$sample_id,meta$File_name)]
spe$tissue_id <- meta$Tissue_ID[match(spe$sample_id,meta$File_name)]
spe$slide_id <- as.factor(meta$Slide_ID[match(spe$sample_id,meta$File_name)])


# add ROI

spe$ROI <- as.vector(str_extract_all(spe$sample_id, "ROI00[1-8]", simplify = TRUE))

unique(spe$timepoint)
```

We have 2 time points:

* Baseline - pre-CAR T cells infusion
* 01MOpost - 1 month post infusion

**Transform counts**

The distribution of expression counts across cells is often observed to be
skewed towards the right side meaning lots of cells display low counts and few
cells have high counts. To avoid analysis biases from these high-expressing
cells, the expression counts are commonly transformed or clipped.

Here, we perform counts transformation using an inverse hyperbolic sine
function. This transformation is commonly applied to [flow cytometry
data](https://support.cytobank.org/hc/en-us/articles/206148057-About-the-Arcsinh-transform).

The `cofactor` here defines the expression range on which no scaling is
performed. While the `cofactor` for CyTOF data is often set to `5`, IMC data
usually display much lower counts. We therefore apply a `cofactor` of `1`.

However, other transformations such as `log(counts(spe) + 0.01)` should be
tested when analysing IMC data.

```{r transform-counts, message=FALSE}
library(dittoSeq)
dittoRidgePlot(spe, var = "CD3", group.by = "timepoint", assay = "counts") +
    ggtitle("CD3 - before transformation")
assay(spe, "exprs") <- asinh(counts(spe)/1)
dittoRidgePlot(spe, var = "CD3", group.by = "timepoint", assay = "exprs") +
    ggtitle("CD3 - inverse hyperbolic sine function")
assay(spe, "logcounts") <- log(counts(spe) + 0.01)
dittoRidgePlot(spe, var = "CD3", group.by = "timepoint", assay = "logcounts") +
    ggtitle("CD3 - logarithm")
```



**Define interesting channels**

Here I set two use_channels, one excluding Histone H3 and membrane markers, one with athose markers usefult to classify the cell types.

```{r select-features}
rowData(spe)$use_channel <- !grepl("ICSK|Histone", rownames(spe))

features_oi <- c('CD45','CD45RO','CD3','CD4','CD8','CD20','CD27','CD138','CD14','CD16','CD56','CD68','CD11c','CD11b','CD31','aSMA')

rowData(spe)$features_oi <- rownames(spe) %in% features_oi
```

**Define color schemes**

We will define color schemes for different metadata entries of the data and
conveniently store them in the `metadata` slot of the `SpatialExperiment` which
will be helpful for downstream data visualizations. We will use colors from the
`RColorBrewer` and `dittoSeq` package but any other coloring package will
suffice.

```{r define-colors, message=FALSE}
library(RColorBrewer)
color_vectors <- list()

ROI <- setNames(brewer.pal(length(unique(spe$ROI)), name = "BrBG"), 
                unique(spe$ROI))
timepoint <- setNames(c('#ff0000', '#0092ff'),
                unique(spe$timepoint))


color_vectors$ROI <- ROI
color_vectors$timepoint <- timepoint

metadata(spe)$color_vectors <- color_vectors
```


## Read in images {#read-images}

The `cytomapper` package allows multi-channel image handling and visualization
within the Bioconductor framework. The most common data format for multi-channel
images or segmentation masks is the TIFF file format, which is used by
`steinbock`.

Here, we will read in multi-channel images and segmentation masks into a
[CytoImageList](https://www.bioconductor.org/packages/release/bioc/vignettes/cytomapper/inst/doc/cytomapper.html#5_The_CytoImageList_object)
data container. It allows storing multiple multi-channel images and requires
matched channels across all images within the object.

The `loadImages` function is used to read in processed multi-channel images and
their corresponding segmentation masks. Of note, the multi-channel images
generated by `steinbock` are saved as 32-bit images while the segmentation masks
are saved as 16-bit images. To correctly scale pixel values of the segmentation
masks when reading them in set `as.is = TRUE`.

```{r read-images}
gc()
images <- loadImages("../data/non_denoised//img/")
gc()

masks <- loadImages("../data/non_denoised//masks/", as.is = TRUE)
gc()
```

In the case of multi-channel images, it is beneficial to set the `channelNames`
for easy visualization. Using the `steinbock` framework, the channel order of
the single-cell data matches the channel order of the multi-channel images.
However, it is recommended to make sure that the channel order is identical
between the single-cell data and the images.

```{r set-channelNames}
channelNames(images) <- rownames(spe)
images
```

For image and mask visualization we will need to
add additional metadata to the `elementMetadata` slot of the `CytoImageList`
objects. This slot is easily accessible using the `mcols` function.

Here, we will save the matched `sample_id`, `patient_id` and `indication`
information within the `elementMetadata` slot of the multi-channel images and
segmentation masks objects. It is crucial that the order of the images in 
both `CytoImageList` objects is the same.

```{r images-dimensions}
df <- data.frame(matrix(ncol = 3, nrow = length(images)))
colnames(df) <- c("sample_id", "width_px", "height_px")

# Fill in the dataframe with object names and feature dimensions
for (i in seq_along(images)) {
  image <- images[[i]]
  df[i, "sample_id"] <- names(images)[i]
  df[i, "width_px"] <- image@dim[1]
  df[i, "height_px"] <- image@dim[2]
}



spe$width_px <- df$width_px[match(spe$sample_id,df$sample_id)]
spe$height_px <- df$height_px[match(spe$sample_id,df$sample_id)]


```


```{r add-metadata}
all.equal(names(images), names(masks))
# adapt to our names
sample_id <- str_split_i(names(images), '_', 2)
timepoint <- meta$Timepoint[match(names(images), meta$File_name)] 

mcols(images) <- mcols(masks) <- DataFrame(file_id = names(images),
                                           sample_id = sample_id,
                                           timepoint = timepoint)
```



# Single-cell visualization

The following section focuses on visualizing the single-cell data contained
in the `SpatialExperiment` object. The main R/Bioconductor packages
to support visualization are `r BiocStyle::Biocpkg("dittoSeq")` and
`r BiocStyle::Biocpkg("imcRtools")`

## Dimensionality reduction


# Dimensionality reduction and visualization

Before doing the batch correction, I will try to run PCA, KNN, UMAP of non-integrated data:

```{r dimred, message=FALSE, cache=TRUE}
set.seed(12345)
spe <- runUMAP(spe, subset_row = rowData(spe)$features_oi, exprs_values = "exprs", BPPARAM=multicore) 
```

After dimensionality reduction, the low-dimensional embeddings are stored in the 
`reducedDim` slot.

```{r show-dimred-slot, cache=TRUE}
reducedDims(spe)
head(reducedDim(spe, "UMAP"))
```



```{r visualizing-dimred-1, fig.height=12, fig.width=8, message=FALSE, paged.print=FALSE, cache=TRUE}
library(patchwork)
library(dittoSeq)
library(viridis)

# visualize patient id 
p1 <- dittoDimPlot(spe, var = "patient_n", reduction.use = "UMAP", size = 0.1, legend.size = 3) + 
    ggtitle("Patient ID on UMAP")
p2 <- dittoDimPlot(spe, var = "slide_id", reduction.use = "UMAP", size = 0.1, legend.size = 3) + 
    ggtitle("Slide ID on UMAP")

# visualize indication
p3 <- dittoDimPlot(spe, var = "timepoint", reduction.use = "UMAP", size = 0.1, legend.size = 3) + 
    scale_color_manual(values = metadata(spe)$color_vectors$timepoint) +
    ggtitle("Timepoint on UMAP")
p4 <- dittoDimPlot(spe, var = "CD3", reduction.use = "UMAP", size = 0.1, legend.size = 3) + 
    ggtitle("CD3 on UMAP")
# visualize indication
p5 <- dittoDimPlot(spe, var = "CD138", reduction.use = "UMAP", size = 0.1, legend.size = 3) + 
    ggtitle("CD138 on UMAP")
p6 <- dittoDimPlot(spe, var = "CD31", reduction.use = "UMAP", size = 0.1, legend.size = 3) + 
    ggtitle("CD31 on UMAP")

(p1 + p2) / (p3 + p4) / (p5+p6)
```



For now not really different results


## Visualizing marker expression 

This section focuses on visualizing the expression of all markers and 
highlighting variation between cells, images and patients.

### Per cell

First, we will visualize single-cell marker expression in form of a
heatmap. Here, we sub-sample the dataset to 2000 cells for visualization
purposes and overlay the cancer type from which the cells were extracted.

```{r segmentation-heatmap, message=FALSE, fig.height=7}
library(ggplot2)
library(viridis)
cur_cells <- sample(seq_len(ncol(spe)), 2000)

dittoHeatmap(spe[,cur_cells], genes = rownames(spe)[rowData(spe)$use_channel],
             assay = "exprs", cluster_cols = TRUE, scale = "none",
             heatmap.colors = viridis(100), annot.by = c("timepoint", "slide_id", "patient_n"),
             annotation_colors = list(timepoint = metadata(spe)$color_vectors$timepoint))
```



### Per image

It can be beneficial to visualize the mean marker expression per image
to identify images with outlying marker expression. This check does not
indicate image quality _per se_ but can highlight biological differences. Here,
we will use the `aggregateAcrossCells` function of the 
`r BiocStyle::Biocpkg("scuttle")` package to compute the mean expression per 
image. For visualization purposes, we again `asinh` transform the mean expression
values.

```{r mean-expression-per-image, message=FALSE, fig.height=7}
library(scuttle)

image_mean <- aggregateAcrossCells(spe, 
                                   ids = spe$sample_id, 
                                   statistics="mean",
                                   use.assay.type = "counts")
assay(image_mean, "exprs") <- asinh(counts(image_mean))

dittoHeatmap(image_mean, genes = rownames(spe)[rowData(spe)$features_oi],
             assay = "exprs", cluster_cols = TRUE, scale = "none",
             heatmap.colors = viridis(100), 
             annot.by = c("timepoint", "slide_id", "patient_n"),
             annotation_colors = list(timepoint = metadata(spe)$color_vectors$timepoint),
             show_colnames = TRUE)
```

### Staining per slide and per sample

We will use ridgeline visualizations to check
differences in staining patterns and biological differences in expression:

```{r ridges, message=FALSE, fig.width=10, fig.height=15}
library(grid)
library(gridExtra)


p <- multi_dittoPlot(spe, vars = c('CD138', 'CD3', 'CD20', 'CD68', 'CD31'),
               group.by = "patient_n", plots = "ridgeplot", 
               assay = "exprs", list.out = TRUE)
             

grid.arrange(grobs = p,nrow=1,top=textGrob("Expression by patient number"))

```




```{r ridges2, message=FALSE, fig.width=10, fig.height=15}
p <- multi_dittoPlot(spe, vars =  c('CD138', 'CD3', 'CD20', 'CD68', 'CD31'),
               group.by = "slide_id", plots = "ridgeplot", 
               assay = "exprs", list.out = TRUE)

grid.arrange(grobs = p,nrow=1,top=textGrob("Expression by slide number"))

```


We observe variations in the distributions of marker expression across patients/slides.
These variations may arise partly from different abundances of cells in
different images as well as staining differences between samples.

## Image area covered by cells

A quality indicator for region selection is the image area covered by cells (or
biological tissue). This metric identifies regions of interest (ROIs) where
little cells are present, possibly hinting at incorrect selection of the ROI. We
can compute the percentage of covered image area using the metadata contained in
the `SpatialExperiment` object:

```{r cell-density, fig.height=6, fig.width=10, message=FALSE}
library(dplyr)

colData(spe) %>%
    as.data.frame() %>%
    group_by(sample_id) %>%
    summarize(cell_area = sum(area),
           no_pixels = mean(width_px) * mean(height_px)) %>%
    mutate(covered_area = cell_area / no_pixels) %>%
    ggplot() +
        geom_point(aes(reorder(sample_id,covered_area), covered_area)) + 
        theme_minimal(base_size = 15) +
        ylim(c(0, 1)) + 
        theme(axis.text.x = element_text(angle = 90, hjust = 1, size = 8)) +
        ylab("% covered area") + xlab("")
```



## Cell size

Next, we observe the distributions of cell size across the individual images.
Differences in cell size distributions can indicate segmentation biases due to
differences in cell density or can indicate biological differences due to cell
type compositions (tumor cells tend to be larger than immune cells). 

```{r cell-size, fig.height=6, fig.width=10, message=FALSE}
colData(spe) %>%
    as.data.frame() %>%
    group_by(sample_id) %>%
    ggplot() +
        geom_boxplot(aes(sample_id, area)) +
        theme_minimal(base_size = 15) + 
        theme(axis.text.x = element_text(angle = 90, hjust = 1, size = 8)) +
        ylab("Cell area") + xlab("")

summary(spe$area)
```

The median cell size is `r median(spe$area)` pixels with a median major axis
length of `r round(median(spe$axis_major_length), digits = 1)`. The largest cell
has an area of `r max(spe$area)` pixels which relates to a diameter of 
`r round(sqrt(max(spe$area)), digits = 1)` pixels assuming a circular shape.

Are the outliers megakacyocytes?


remove too small cells:

```{r remove-small-cells}
sum(spe$area < 5)
spe <- spe[,spe$area >= 5]
```



## Save objects

Finally, the generated data objects can be saved for further downstream 
processing and analysis.

```{r save-objects-read-data}
rm(list=setdiff(ls(), c('images', 'spe','masks')))

invisible(gc())
saveRDS(spe, "../data/non_denoised/spe.rds")
rm(spe)
invisible(gc())
saveRDS(images, "../data/non_denoised/images.rds")
rm(images)
invisible(gc())
saveRDS(masks, "../data/non_denoised/masks.rds")
```


## Session info

```{r}
sessionInfo()
```
