==> notebooks/10-Import_to_python.ipynb <==
import squidpy as sq
import scanpy as sc
import scanpy as sc
import anndata as ad
import os
import sys
import pandas as pd
import numpy as np
import tifffile as tf

adats_list = os.listdir('../data/non_denoised/anndata/')
img_list = os.listdir('../data/non_denoised/img/')
masks_list = os.listdir('../data/non_denoised/masks/')

adatas=[]
imgs=[]

def preprocess_adata(sample):
    sample = sample.split('.')[0]
    filepath= '../data/non_denoised/anndata/' + sample + '.h5ad'
    imgpath='../data/non_denoised/img/' + sample + '.tiff'
    maskpath='../data/non_denoised/masks/' + sample + '.tiff'
    library_id = sample
    sample_id = 's'+sample.split('_')[1]
    adata=sc.read(filepath)
    new_obs_names = [sample_id +'_'+ object.split()[1] for object in adata.obs_names]
    adata.obs['Cell_ID'] = adata.obs_names.str.split().str[1].astype('int')
    adata.obs_names = new_obs_names
    adata.obs['library_id']=library_id
    coordinates = pd.concat([adata.obs['centroid-0'],adata.obs['centroid-1']],axis=1).to_numpy()
    adata.obsm['spatial'] = coordinates
    img=np.transpose(tf.imread(imgpath), (1, 2, 0))
    mask=tf.imread(maskpath)
    uns_dict = {
        'spatial':{
            library_id:{
                'metadata':{
                    'source_image_path':imgpath,
                    'source_mask_path':maskpath
                },
                'images':{
                    'hires':img,
                    'segmentation':mask
                },
                'scalefactor':{
                    'tissue_hires_scalef': 1,
                    'spot_diameter_fullres': 1
                }
            }
        }
    }
    adata.uns = uns_dict
    return adata, library_id


def preprocess_images(adata):
    library_id = list(adata.uns['spatial'].keys())[0]
    img_path=adata.uns['spatial'][library_id]['metadata']['source_image_path']
    img = sq.im.ImageContainer(adata.uns['spatial'][library_id]['metadata']['source_image_path'],
                               layer='img',
                               lazy=False)
    img.add_img(adata.uns['spatial'][library_id]['metadata']['source_mask_path'],
               layer='mask',
               lazy=False)
    return img
    
    
    
# for sample in adats_list:
#     adatas.append(preprocess_adata(sample))
#     adatas[sample_id] = adata

adatas=[]
imgs=[]
library_ids = []
for sample in adats_list:
    adata, library_id = preprocess_adata(sample)
    library_ids.append(library_id)
    adatas.append(adata)
    imgs.append(preprocess_images(adatas[-1]))
    

adata = ad.concat(
    adatas, uns_merge="only")


adata=adata[adata.obs['area']>5]

def read_reindex(file):
    df=pd.read_csv('../data/non_denoised/csv_tables/'+file+'.csv', index_col=[0])
    df=df.reindex(adata.obs_names)
    return df

exprs=read_reindex('exprs')

counts=read_reindex('counts')
logcounts=read_reindex('logcounts')

observations_metadata=read_reindex('observations_metadata')

fastMNN=read_reindex('fastMNN')
UMAP_mnnCorrected=read_reindex('UMAP_mnnCorrected')
PCA=read_reindex('PCA')

features_metadata=pd.read_csv('../data/non_denoised/csv_tables/features_metadata.csv', index_col=[0])

adata.X = counts.to_numpy()
adata.layers['exprs'] = exprs.to_numpy()
adata.layers['logcounts'] = logcounts.to_numpy()

adata.obs

adata.obs.columns.tolist()

observations_metadata=pd.concat([adata.obs,observations_metadata.loc[:,~observations_metadata.columns.isin(adata.obs.columns.tolist())]],ignore_index=False,axis=1)
adata.obs = observations_metadata

adata.var_names = features_metadata.index
adata.var = features_metadata

adata.obsm['PCA'] = PCA.to_numpy()
adata.obsm['fastMNN'] = fastMNN.to_numpy()
adata.obsm['UMAP_mnnCorrected'] = UMAP_mnnCorrected.to_numpy()


adata.write('../data/non_denoised/spe_minor_celltypes.h5ad')

adata.obs

import pickle

# Open a file and use dump() 
with open('../data/non_denoised/images.pkl', 'wb') as file: 
    # A new file will be created 
    pickle.dump(imgs, file)
